int to float conversion

+++ DG_backend/app/models/deal.py
@@ -3,7 +3,7 @@ from pydantic import BaseModel
 
 class Deal(Document):
     title: str          # e.g., "Deal 1"
-    price: int          # e.g., 480
+    price: float        # e.g., 480.0
     description: str    # e.g., "500g Chicken pulao..."
     image_url: str
